refactor(optimize): unify dense/sparse eval loops into single evaluator
Extract `evaluate_site_contributions()` in `optimize_eval.rs` taking
`impl Iterator<Item = (f64, ArrayView1<f64>)>` for (multiplicity,
coefficients). Dense wraps rows as `(1.0, row)`, sparse passes
`(multiplicity, coefficients.view())`. The 4 duplicated inner loops
collapse to 2 (with/without derivatives), each written once.

Thin wrappers in `optimize_dense_eval.rs` and `optimize_sparse_eval.rs`
preserve existing call sites. Update known issue to reflect remaining
enum dispatch duplication.
