+++ packages/treetime/src/commands/optimize/mod.rs
@@ -4,6 +4,7 @@ mod __tests__;
 pub mod args;
 pub mod optimize_dense;
 mod optimize_dense_eval;
+mod optimize_eval;
 pub mod optimize_indel;
 pub mod optimize_sparse;
 mod optimize_sparse_eval;
